fix: 없는 여행, 숙박 Id 에 대한 에러 처리

+++ src/main/java/org/example/kdtbe8_toyproject2/accommodation/service/AccommodationService.java
@@ -49,6 +49,9 @@ public class AccommodationService {
 
     public void delete(Long id, Long tripId) {
         var accommodationEntity = accommodationMapper.findByTripId(tripId);
+        if (accommodationEntity == null || accommodationEntity.isEmpty()) {
+            throw AccommodationError.TRIP_NOT_EXIST.defaultException();
+        }
         if (accommodationMapper.delete(id) == 0) {
             throw AccommodationError.ACCOMMODATION_NOT_EXIST.defaultException();
